+++ baseclient.go
@@ -13,7 +13,7 @@ import (
 	"github.com/google/go-querystring/query"
 )
 
-const sdkVersion = "7.6.1"
+const sdkVersion = "7.11.1"
 
 const lookupBaseUrl = "lookup.plivo.com"